+++ src/goose/__init__.py
@@ -1,4 +1,5 @@
 from .pluginbase.notifier import Notifier  # noqa: F401
 from .pluginbase.toolkit import Toolkit, tool  # noqa: F401
 from .pluginbase.profile import Profile, ToolkitSpec  # noqa: F401
+from .pluginbase.command import Command  # noqa: F401
 from .pluginbase import utils  # noqa: F401
