+++ src/goose/pluginbase/command.py
