+++ tests/_internal/cli/prompt/test_completer.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock
 
 import pytest
 from goose._internal.cli.prompt.completer import GoosePromptCompleter
-from goose.command.base import Command
+from goose.pluginbase.command import Command
 from prompt_toolkit.completion import Completion
 from prompt_toolkit.document import Document
